// src/Types/postsDataTypes.ts
export interface PostDetails {
  _id: string;
  category: string;
  action: string;
  Country: string;
  address: string;
  propertyType: string;
  Space: number;
  SalePrice: number;
  RentPrice: number;
  description: string;
  Name: string;
  phone: string;
  images: string[];
  selectedDates: string[];
  __v: number;
}